Move tagged versions into master
Merged all standalone branches with tarantool tags
into master branch to be able to build the tarantool
tagged versions in master and remove branches copies.

Found that 2.2 build failed on SSL, to fix it the
dockerfile changed from 2.2 to 2.x to use openssl
as described in #122.

Also removed curl in 2.2.1 tagged dockerfile as suggested
in #122, after which it became common to 2.2 dockerfile,
so the file was removed as duplicating and all jobs builds
moved from 2.2.1 to 2.2 dockerfile.