Change the dockerfile for Tarantool 2.1
Before bumping alpine from 3.5 to 3.9 at commit:
  aef661e81bc8e7c26f4ac6f516d83a69204653b3 "Bump alpine images version"

Tarantool versions at 2.1 branch with versions equal and later than
2.1.3 tag failed building on alpine 3.5 with both dockerfiles
that were available for it, due to issue:

  /usr/src/tarantool/third_party/curl/lib/vtls/openssl.c: In function 'set_ssl_version_min_max':
  /usr/src/tarantool/third_party/curl/lib/vtls/openssl.c:2195:9: error: implicit declaration of function 'SSL_CTX_set_min_proto_version' [-Werror=implicit-function-declaration]
       if(!SSL_CTX_set_min_proto_version(ctx, ossl_ssl_version_min)) {
           ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~

It happened because, before the mentioned commit this gitlab-ci job
used for building 'dockerfile/alpine_3.5_2.2', which was renamed to
'dockerfile/alpine_3.5_2.x' at that commit, but the job's variables
'DIST' and 'VER' were not completely updated. So the gitlab-ci job
build was broken after that commit, due to not found dockerfile.

Testing manually the current job with the currently nested for it
'dockerfile/alpine_3.5_2.x' produced the shown issue. It occurred
because of the unsupported openssl version, please check [1].
These macros are only available in OpenSSL 1.1.0. While alpine
versions uses different openssl versions:
Alpine 3.5: openssl (1.0.2q-r0)
Alpine 3.9: openssl (1.1.1g-r0)

Also in #152 the alpine version was bumped and current change of the
build job for Tarantool 2.1 to use alpine 3.9 resolved the issue.

Currently found that on alpine 3.9 it successfully built, but only
with *1.x dockerfile, due to tarantoolctl tool failed installing rocks
at *2.x file.

So decided to use for Tarantool 2.1 -> alpine_3.9_1.x dockerfile.

Closes #173
Follows up #152

[1] - https://www.openssl.org/docs/man1.1.0/ssl/SSL_CTX_set_max_proto_version.html