readme: update the release policy
Until a recent discussion we have to update existing docker tag on
Docker Hub if something is fixed in a Dockerfile. Now we reconsidered
the process and decided to never change existing docker tags.

The reason is simple: there should be a way to freeze an image to get
rid of all possible negative impact due to a change.

The discussion is raised by @tsafin. Thank you, Timur!